WIP to improve the budget editing to use AJAX techniques.
Work in progress of improving the budget editing to use AJAX techniques.
Rather than reloading the entire tree, I should be reloading an
individual node as it is edited / committed.  I started this by adding
support for it in the backend; the GUI is not there yet (and is actually
currently broken completely).